stash draft changes migrating to bedrock.ts instead of bedrock-chat new, before removing deprecated module and looking into stream interface
