+++ docs/extras/modules/model_io/models/llms/integrations/bedrock.mdx
@@ -8,7 +8,7 @@ from leading AI startups and Amazon available via an API. You can choose from a 
 You'll need to install a few official AWS packages as peer dependencies:
 
 ```bash npm2yarn
-npm install @aws-crypto/sha256-js @aws-sdk/credential-provider-node @aws-sdk/protocol-http @aws-sdk/signature-v4
+npm install @aws-crypto/sha256-js @aws-sdk/credential-provider-node @aws-sdk/protocol-http @aws-sdk/signature-v4 @aws-sdk/eventstream-marshaller @aws-sdk/util-utf8-universal
 ```
 
 ## Usage
@@ -17,20 +17,3 @@ import CodeBlock from "@theme/CodeBlock";
 import BedrockExample from "@examples/models/llm/bedrock.ts";
 
 <CodeBlock language="typescript">{BedrockExample}</CodeBlock>
-
-# Bedrock Chat
-
-## Setup
-
-You'll need to install a few official AWS packages as peer dependencies:
-
-```bash npm2yarn
-npm install @aws-crypto/sha256-js @aws-sdk/credential-provider-node @aws-sdk/protocol-http @aws-sdk/signature-v4 @aws-sdk/eventstream-marshaller @aws-sdk/util-utf8-universal
-```
-
-## Usage
-
-import CodeBlock from "@theme/CodeBlock";
-import BedrockExample from "@examples/models/llm/bedrock-chat.ts";
-
-<CodeBlock language="typescript">{BedrockChatExample}</CodeBlock>
