+++ langchain/.gitignore
@@ -133,9 +133,6 @@ llms/sagemaker_endpoint.d.ts
 llms/bedrock.cjs
 llms/bedrock.js
 llms/bedrock.d.ts
-llms/bedrock-chat.cjs
-llms/bedrock-chat.js
-llms/bedrock-chat.d.ts
 prompts.cjs
 prompts.js
 prompts.d.ts
