+++ langchain/scripts/create-entrypoints.js
@@ -59,7 +59,6 @@ const entrypoints = {
   "llms/googlepalm": "llms/googlepalm",
   "llms/sagemaker_endpoint": "llms/sagemaker_endpoint",
   "llms/bedrock": "llms/bedrock",
-  "llms/bedrock-chat": "llms/bedrock-chat",
   // prompts
   prompts: "prompts/index",
   "prompts/load": "prompts/load",
@@ -263,8 +262,7 @@ const requiresOptionalDependency = [
   "llms/raycast",
   "llms/replicate",
   "llms/sagemaker_endpoint",
-  "llms/bedrock",  
-  "llms/bedrock-chat",
+  "llms/bedrock",
   "prompts/load",
   "vectorstores/analyticdb",
   "vectorstores/chroma",
