+++ langchain/src/llms/tests/bedrock.int.test.ts
@@ -1,5 +1,8 @@
 import { test, expect } from "@jest/globals";
+import { EventStreamMarshaller } from "@aws-sdk/eventstream-marshaller";
+import { fromUtf8, toUtf8 } from "@aws-sdk/util-utf8-universal";
 import { Bedrock } from "../bedrock.js";
+import { CallbackManager } from "../../callbacks/index.js";
 
 test("Test Bedrock LLM: ai21", async () => {
   const region = "us-east-1";
@@ -17,7 +20,7 @@ test("Test Bedrock LLM: ai21", async () => {
     ): Promise<Response> {
       expect(input).toBeInstanceOf(URL);
       expect((input as URL).href).toBe(
-        `https://bedrock.${region}.amazonaws.com/model/${model}/invoke`
+        `https://bedrock.${region}.amazonaws.com/model/${model}/invoke-with-response-stream`
       );
       expect(init?.method).toBe("POST");
       expect(init?.headers).toMatchObject({
@@ -26,13 +29,14 @@ test("Test Bedrock LLM: ai21", async () => {
         "Content-Type": "application/json",
       });
       expect(init?.body).toBe(`{"prompt":"${prompt}"}`);
-      return new Promise<Response>((resolve) => {
-        resolve(
-          new Response(`{"completions":[{"data":{"text":"${answer}"}}]}`, {
-            status: 200,
-            headers: { "Content-Type": "application/json" },
-          })
-        );
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      return new Promise<any>((resolve) => {
+        resolve({
+          status: 200,
+          body: {
+            getReader: () => buildResponse(answer, "data.text"),
+          },
+        });
       });
     },
   });
@@ -58,7 +62,7 @@ test("Test Bedrock LLM: anthropic", async () => {
     ): Promise<Response> {
       expect(input).toBeInstanceOf(URL);
       expect((input as URL).href).toBe(
-        `https://bedrock.${region}.amazonaws.com/model/${model}/invoke`
+        `https://bedrock.${region}.amazonaws.com/model/${model}/invoke-with-response-stream`
       );
       expect(init?.method).toBe("POST");
       expect(init?.headers).toMatchObject({
@@ -69,13 +73,14 @@ test("Test Bedrock LLM: anthropic", async () => {
       expect(init?.body).toBe(
         `{"prompt":"${prompt}","max_tokens_to_sample":50}`
       );
-      return new Promise<Response>((resolve) => {
-        resolve(
-          new Response(`{"completion":"${answer}"}`, {
-            status: 200,
-            headers: { "Content-Type": "application/json" },
-          })
-        );
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      return new Promise<any>((resolve) => {
+        resolve({
+          status: 200,
+          body: {
+            getReader: () => buildResponse(answer, "completion"),
+          },
+        });
       });
     },
   });
@@ -101,7 +106,7 @@ test("Test Bedrock LLM: amazon", async () => {
     ): Promise<Response> {
       expect(input).toBeInstanceOf(URL);
       expect((input as URL).href).toBe(
-        `https://bedrock.${region}.amazonaws.com/model/${model}/invoke`
+        `https://bedrock.${region}.amazonaws.com/model/${model}/invoke-with-response-stream`
       );
       expect(init?.method).toBe("POST");
       expect(init?.headers).toMatchObject({
@@ -112,13 +117,14 @@ test("Test Bedrock LLM: amazon", async () => {
       expect(init?.body).toBe(
         '{"inputText":"What is your name?","textGenerationConfig":{}}'
       );
-      return new Promise<Response>((resolve) => {
-        resolve(
-          new Response(`{"results":[{"outputText":"${answer}"}]}`, {
-            status: 200,
-            headers: { "Content-Type": "application/json" },
-          })
-        );
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      return new Promise<any>((resolve) => {
+        resolve({
+          status: 200,
+          body: {
+            getReader: () => buildResponse(answer, "outputText"),
+          },
+        });
       });
     },
   });
@@ -171,3 +177,103 @@ test("Test Bedrock LLM: no-region-specified", async () => {
     "Unknown model: 'other.model', only these are supported: ai21,anthropic,amazon"
   );
 }, 5000);
+
+test("Test Bedrock LLM: streaming", async () => {
+  const region = "us-east-1";
+  const model = "amazon.model";
+  const prompt = "What is your name?";
+  const answer = "Hello! My name is Claude.";
+
+  const tokens: string[] = [];
+  const callbackManager = CallbackManager.fromHandlers({
+    async handleLLMNewToken(token: string) {
+      tokens.push(token);
+    },
+  });
+
+  const bedrock = new Bedrock({
+    maxTokens: 20,
+    region,
+    model,
+    async fetchFn(
+      input: RequestInfo | URL,
+      init?: RequestInit | undefined
+    ): Promise<Response> {
+      expect(input).toBeInstanceOf(URL);
+      expect((input as URL).href).toBe(
+        `https://bedrock.${region}.amazonaws.com/model/${model}/invoke-with-response-stream`
+      );
+      expect(init?.method).toBe("POST");
+      expect(init?.headers).toMatchObject({
+        host: `bedrock.${region}.amazonaws.com`,
+        accept: "application/json",
+        "Content-Type": "application/json",
+      });
+      expect(init?.body).toBe(
+        `{"inputText":"${prompt}","textGenerationConfig":{}}`
+      );
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      return new Promise<any>((resolve) => {
+        resolve({
+          status: 200,
+          body: {
+            getReader: () => buildResponse(answer, "outputText"),
+          },
+        });
+      });
+    },
+    callbackManager,
+  });
+
+  const res = await bedrock.call(prompt);
+  expect(typeof res).toBe("string");
+  expect(res).toBe(answer);
+  expect(tokens.join("")).toBe(answer);
+}, 5000);
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+function setValue(object: any, path: string, value: any) {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any no-useless-escape
+  const pathN: string = path.replace(/[\[]/gm, ".").replace(/[\]]/gm, "");
+  const keys: string[] = pathN.split(".");
+  const last: string | undefined = keys.pop();
+  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion no-param-reassign no-return-assign
+  keys.reduce((o, k) => (o[k] = o[k] || {}), object)[last!] = value;
+
+  return object;
+}
+
+function buildResponse(tokens: string, keys = "outputText") {
+  const body = setValue({}, keys, tokens);
+  const bytes = JSON.stringify({
+    bytes: Buffer.from(JSON.stringify(body)).toString("base64"),
+  });
+  const event = new EventStreamMarshaller(toUtf8, fromUtf8).marshall({
+    headers: {
+      ":event-type": { type: "string", value: "chunk" },
+      ":content-type": { type: "string", value: "application/json" },
+      ":message-type": { type: "string", value: "event" },
+    },
+    body: Uint8Array.from(Buffer.from(bytes)),
+  });
+
+  let chunkIter = 0;
+  const mockReader = {
+    read: async () => {
+      if (chunkIter === 0) {
+        chunkIter += 1;
+        return {
+          done: false,
+          value: event,
+        };
+      } else {
+        chunkIter += 1;
+        return {
+          done: true,
+        };
+      }
+    },
+  };
+
+  return mockReader;
+}
