+++ langchain/tsconfig.json
@@ -75,7 +75,6 @@
       "src/llms/googlepalm.ts",
       "src/llms/sagemaker_endpoint.ts",
       "src/llms/bedrock.ts",
-      "src/llms/bedrock-chat.ts",
       "src/prompts/index.ts",
       "src/prompts/load.ts",
       "src/vectorstores/analyticdb.ts",
